// src/interfaces/User.ts
export interface UserResponse {
  login: string;
  id: number;
  node_id: string;
  avatar_url: string;
  gravatar_id: string;
  url: string;
  html_url: string;
  followers_url: string;
  following_url: string;
  gists_url: string;
  starred_url: string;
  subscriptions_url: string;
  organizations_url: string;
  repos_url: string;
  events_url: string;
  received_events_url: string;
  type: string;
  site_admin: boolean;
  name: string;
  company: string;
  blog: string;
  location: string;
  email: string | null;
  hireable: string | null;
  bio: string | null;
  twitter_username: null;
  public_repos: number;
  public_gists: number;
  followers: number;
  following: number;
  created_at: string;
  updated_at: string;
}

export interface User{
    name: string;
    login: string;
    avatar_url: string;
    created_at: string;
    public_repos: number;
    followers: number;
    following: number;  
    twitter_username: string | null;
    bio: string | null;
    blog: string | null;
    company: string | null;
    location: string | null;
}